typed-protocols:codec-properties library
Moved `prop_codec` properties to a sublibrary.  All properties return
QuickCheck's `Property` type.
