+++ typed-protocols-doc/typed-protocols-doc.cabal
@@ -89,7 +89,7 @@ test-suite typed-protocols-doc-test
     build-depends: base >=4.14.0.0 && <5
                  , blaze-html >=0.9.1.2 && <0.10
                  , tasty >=1.5 && <1.6
-                 , tasty-quickcheck >=0.10.3 && <0.11
+                 , tasty-quickcheck >=0.10.3 && <0.12
                  , typed-protocols
                  , typed-protocols-doc
                  , serdoc-core
